refactor(scanner): Extract VCSPath filtering functions
Scanners generally do not consider the VCSPath when performing an
ORT scan on a repository, as they scan the entire repository and
its dependencies. The resulting scan data is cached for reuse
across multiple scans.

When reusing cached scan results for ORT scans with a specified
VCSPath, where only a subset of the repository is scanned, it
is necessary to filter out extraneous scan results. Helper
functions already exist to handle this filtering.

This refactoring extracts these helper functions and allows
ORT Server to use them.
See: https://github.com/eclipse-apoapsis/ort-server/pull/1358
